+++ packages/angular-query-experimental/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/angular-query-experimental
 
+## 5.102.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.102.8
+
+## 5.102.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.102.7
+
+## 5.102.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.102.6
+
 ## 5.102.5
 
 ### Patch Changes