+++ packages/query-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tanstack/query-core
 
+## 5.102.8
+
+## 5.102.7
+
+## 5.102.6
+
 ## 5.102.5
 
 ### Patch Changes