+++ packages/vue-query-devtools/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/vue-query-devtools
 
+## 6.1.48
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.102.8
+  - @tanstack/vue-query@5.102.8
+
+## 6.1.47
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.102.7
+  - @tanstack/vue-query@5.102.7
+
+## 6.1.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.102.6
+  - @tanstack/vue-query@5.102.6
+
 ## 6.1.45
 
 ### Patch Changes